requirements.txt
Django==4.2.7
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
whitenoise==6.5.0
gunicorn==21.2.0 